changes :  - fix for the screen reload fix on alert message  - https://github.com/omgnetwork/optimism/issues/355
