+++ packages/omgx/wallet-frontend/src/layout/index.js
@@ -17,19 +17,16 @@ import { Box, useMediaQuery } from '@material-ui/core';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { createTheme, responsiveFontSizes, ThemeProvider } from '@material-ui/core/styles';
 import { setWalletMethod } from 'actions/setupAction';
-import { closeAlert, closeError } from 'actions/uiAction';
-import Alert from 'components/alert/Alert';
 import WalletPicker from 'components/walletpicker/WalletPicker';
 import Home from 'containers/home/Home';
 import Notification from 'containers/notification/Notification';
 import React, { Suspense, useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import {
   BrowserRouter as Router,
   Route,
   Switch
 } from "react-router-dom";
-import { selectAlert, selectError } from 'selectors/uiSelector';
 import { isChangingChain } from 'util/changeChain';
 //import oracleService from 'services/oracleService';
 import * as styles from './layout.module.scss';
@@ -40,14 +37,8 @@ function App () {
   const dispatch = useDispatch();
   const themeFromLocalStorage = localStorage.getItem('theme');
 
-  const errorMessage = useSelector(selectError);
-  const alertMessage = useSelector(selectAlert);
-
   const [ enabled, setEnabled ] = useState(false);
   const [light, setLight] = useState(themeFromLocalStorage === 'light');
-
-  const handleErrorClose=()=>dispatch(closeError());
-  const handleAlertClose=()=>dispatch(closeAlert());
 
   let theme = createTheme({
     palette: {
@@ -232,26 +223,6 @@ function App () {
         <Box sx={{ display: 'flex', flexDirection: isMobile ? 'column' : 'row' }}>
           <S.Content>
             <div className={styles.App}>
-
-              <Alert
-                type='error'
-                duration={0}
-                open={!!errorMessage}
-                onClose={handleErrorClose}
-                position={50}
-              >
-                {errorMessage}
-              </Alert>
-
-              <Alert
-                type='success'
-                duration={0}
-                open={!!alertMessage}
-                onClose={handleAlertClose}
-                position={0}
-              >
-                {alertMessage}
-              </Alert>
 
               <Notification/>
               <Suspense fallback={<>Loading...</>}>
